Refactor settings_widgets into more smaller files
