+++ .gitignore
@@ -21,6 +21,7 @@ migrate_working_dir/
 # VS Code which you may wish to be included in version control, so this line
 # is commented out by default.
 .vscode/
+.vs/
 
 # Flutter/Dart/Pub related
 **/doc/api/
