+++ lib/page/settings_page/setting_confirmation.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+
+/// Holds information for showing a confirmation dialog.
+class SettingConfirmation {
+  /// The confirmation title.
+  final String title;
+  /// The confirmation content.
+  final String content;
+  
+
+  SettingConfirmation({required this.title, required this.content});
+}
+
+Future<bool> showConfirmationDialog(BuildContext context, SettingConfirmation confirmation) {
+  return showPlatformDialog(
+    context: context,
+    builder: (context) => PlatformAlertDialog(
+      title: Text(confirmation.title),
+      content: Text(confirmation.content),
+      actions: [
+        PlatformDialogAction(
+          child: PlatformText("Cancel"),
+          onPressed: () {
+            Navigator.of(context).pop(false);
+          },
+        ),
+        PlatformDialogAction(
+          child: PlatformText("Ok"),
+          onPressed: () {
+            Navigator.of(context).pop(true);
+          },
+        ),
+      ],
+    ),
+  ) as Future<bool>;
+}
